fix: properly handle COM lib errors for cloaking
One common source for crashes when using the COM library cloaking is the
get_view_for_hwnd function from IApplicationViewCollection. For a lot of
windows, getting a view is not possible (Outlook splash screen, OneDrive
quit screen, etc). In the situations where the view can't be acquired,
windows are essentially ignored.