fix: avoid recentering manually placed floating windows when unplugging USBs (#1121)
Currently in GlazeWM, it assumes any change reported by the OS is a
display setting change, but sometimes this is just an unrelated event,
such a usb being unplugged/replugged.

The problem is that if the user has a floating window that has been
positioned (either manually moved or moved via the position command),
this window will be automatically recentered because the positioning
might have changed.

This PR checks if the window was placed by the user and if it still
overlaps with the workspace after the change before forcefully
recentering.

---------

Co-authored-by: Lars Berger <lars.berger@yahoo.com>